# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "sms-app api documentation",
        "description": "sms-app api documentation",
        "contact": {
            "email": "orutu1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/configurations": {
            "get": {
                "tags": [
                    "Show All Configuration Settings"
                ],
                "description": "Show All Configuration Settings",
                "operationId": "AllConfigurationSettings",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailConfiguration"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry no contact found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Create Email Configuration Settings"
                ],
                "description": "Create new email configuration settings",
                "operationId": "CreateEmailConfiguration",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "query",
                        "description": "Enter email driver",
                        "required": true
                    },
                    {
                        "name": "host",
                        "in": "query",
                        "description": "Enter email host",
                        "required": true
                    },
                    {
                        "name": "port",
                        "in": "query",
                        "description": "Enter email port",
                        "required": true
                    },
                    {
                        "name": "encryption",
                        "in": "query",
                        "description": "Enter email enycryption ",
                        "required": true
                    },
                    {
                        "name": "user_name",
                        "in": "query",
                        "description": "Enter email user name",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter email password",
                        "required": true
                    },
                    {
                        "name": "sender_name",
                        "in": "query",
                        "description": "Enter email sender name",
                        "required": true
                    },
                    {
                        "name": "sender_email",
                        "in": "query",
                        "description": "Enter email sender email",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration created successfully"
                    },
                    "400": {
                        "description": "Sorry unable to create configuration"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/configurations/{id}": {
            "get": {
                "tags": [
                    "Show Configuration"
                ],
                "description": "Show configuration details",
                "operationId": "ShowConfiguration",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter configuration id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Sorry this data do not exist"
                    }
                }
            },
            "post": {
                "tags": [
                    "Update Email Configuration Settings"
                ],
                "description": "Update email configuration settings",
                "operationId": "UpdateEmailConfiguration",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter configuration id",
                        "required": true
                    },
                    {
                        "name": "driver",
                        "in": "query",
                        "description": "Enter email driver",
                        "required": true
                    },
                    {
                        "name": "host",
                        "in": "query",
                        "description": "Enter email host",
                        "required": true
                    },
                    {
                        "name": "port",
                        "in": "query",
                        "description": "Enter email port",
                        "required": true
                    },
                    {
                        "name": "encryption",
                        "in": "query",
                        "description": "Enter email enycryption ",
                        "required": true
                    },
                    {
                        "name": "user_name",
                        "in": "query",
                        "description": "Enter email user name",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter email password",
                        "required": true
                    },
                    {
                        "name": "sender_name",
                        "in": "query",
                        "description": "Enter email sender name",
                        "required": true
                    },
                    {
                        "name": "sender_email",
                        "in": "query",
                        "description": "Enter email sender email",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration updated successfully"
                    },
                    "400": {
                        "description": "Sorry unable to update configuration"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete Settings"
                ],
                "summary": "Delete existing settings",
                "description": "Deletes configuration settings and returns no content",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry this do not exist"
                    },
                    "400": {
                        "description": "Sorry unable to delete settings"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmailConfiguration": {
                "title": "EmailConfiguration",
                "description": "EmailConfiguration model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "driver": {
                        "title": "Driver",
                        "description": "Enter email driver",
                        "type": "string",
                        "example": "smtp"
                    },
                    "host": {
                        "title": "Host",
                        "description": "Enter email host",
                        "type": "string",
                        "example": "smtp.gmail.com"
                    },
                    "port": {
                        "title": "Port",
                        "description": "Enter email port",
                        "type": "string",
                        "example": "465"
                    },
                    "encryption": {
                        "title": "Encryption",
                        "description": "Enter email encyription",
                        "type": "string",
                        "example": "ssl"
                    },
                    "user_name": {
                        "title": "Username",
                        "description": "Enter email username",
                        "type": "string",
                        "example": "orutu1@gmail.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Enter email password",
                        "type": "string",
                        "example": "password"
                    },
                    "sender_name": {
                        "title": "Sender Name",
                        "description": "Enter email sender name",
                        "type": "string",
                        "example": "AgroTech"
                    },
                    "sender_email": {
                        "title": "Sender Email",
                        "description": "Enter sender email",
                        "type": "string",
                        "example": "orutu1@gmail.com"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EmailConfiguration"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}